Fix KUL packages according to changes in the git repos
There is now a single repository 'kin-gen' that contains the
ilk-generator and ilk-compiler as submodules. Thus, we only
need to integrate 'kin-gen' in the esrocos build system.

The branch of the project is set to 'autoproj', as that is the
one containing the required CMakeLists and manifest.

We removed 'luarocks' from the list associated with the 'lua'
dependency, since luarocks should be installed by hand; the package
manager version might not match the Lua interpreter version.